Bug fixes
1) Removed a bug where the metadata update was executed twice, soemtimes
truncating the _extractors without reason.

2) Removed a bug where, if an extractor that does not take parameters
was specified in the JSON configuration file with a string, it would be
automatically perceived as different as its previous execution and
removed and re-included. This was especially bad, since the SQLCoreUtils
is now added as the first extractor and it would always be perceived as
different, thus making all eecutions as the first one

3) Removed a slight bug in the AnnotationExtractor